Update jquery.jsonview.css
Removed the sourceMappingURL since no .map file exists (and this file isn't minified) for this css and this was causing an unnecessary 404 when opening dev tools: "GET https://JENKINS/plugin/extended-choice-parameter/css/jquery.jsonview.css.map 404 (Not Found)".

Also fixed the transition to be a single line so it isn't overwritten.  Removed a redundant 'px'.  The indent was a little off at the start of all the lines (except the first few), fixed that.